requirements.txt
Django==4.2.7
python-dotenv==1.0.0
django-crispy-forms==2.0
crispy-bootstrap5==0.7
psycopg2-binary==2.9.7
openpyxl==3.1.2
reportlab==4.0.4
matplotlib==3.7.1
seaborn==0.12.2
django-filter==23.3
django-guardian==2.4.0
django-import-export==3.3.0
djangorestframework==3.14.0
celery==5.3.1
redis==4.6.0
django-celery-results==2.5.1